Add path and data to Session
Instead of keeping a special session object inside of the main
application, I moved the session path to the Session object and
moved the payload portion to a subobject, data, within the Session.
This keeps the ephemeral parts of the Session, e.g. the path,
seperate from the parts that should be written to the session.json
file upon save.