Create the Shape object for storing shapes
We need to store the tool type as a string along size the actual
object so as to make reconstruction possible when we send the
shapes to the main processs.

Shape returns an object storing the tool type and the constructed
tool object with prototype methods, add_point and draw, which
forward to the tool object.