Compute timeseries from shapes
We have a performance problem which needs to be addressed in the
near future. Namely, the undo/redo functions are very laggy when
we have a lot a features selected. This seems to be caused by
having to recompute the timeseries despite the fact that only
one shape changes at a time.